:recycle: Refactor MonsterCard in its own file
