+++ source/YugiAcademy/YuGiOh/models.py
@@ -46,15 +46,3 @@ class Card(models.Model):
 
     class Meta:
         abstract = True
-
-
-class MonsterCard(Card):
-    Attributes = models.TextChoices('Attributes', 'Dark Divine Earth Fire Light Water Wind')
-    race = models.CharField(max_length=20)
-    attribute = models.CharField(max_length=10, choices=Attributes.choices)
-    level = models.IntegerField()
-    attack = models.IntegerField()
-    defense = models.IntegerField()
-
-    def related_query_name(self):
-        return 'monster_card'
